requirements.txt
aiomax>=0.1.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
sqlite3